+++ noodles-sam/src/record/builder.rs
@@ -311,7 +311,7 @@ impl Builder {
     ///
     /// let data = Data::try_from(vec![data::Field::new(
     ///     data::field::Tag::AlignmentHitCount,
-    ///     data::field::Value::Int(1),
+    ///     data::field::Value::Int32(1),
     /// )])?;
     /// let record = sam::Record::builder().set_data(data).build()?;
     /// assert_eq!(record.data().to_string(), "NH:i:1");
@@ -433,7 +433,7 @@ mod tests {
 
         let data = Data::try_from(vec![data::Field::new(
             data::field::Tag::AlignmentHitCount,
-            data::field::Value::Int(1),
+            data::field::Value::Int32(1),
         )])?;
 
         let record = Builder::default()