+++ noodles-sam/src/record/data/field.rs
@@ -23,7 +23,7 @@ impl Field {
     ///
     /// ```
     /// use noodles_sam::record::data::{field::{Tag, Value}, Field};
-    /// let field = Field::new(Tag::AlignmentHitCount, Value::Int(1));
+    /// let field = Field::new(Tag::AlignmentHitCount, Value::Int32(1));
     /// ```
     pub fn new(tag: Tag, value: Value) -> Self {
         Self { tag, value }
@@ -35,7 +35,7 @@ impl Field {
     ///
     /// ```
     /// use noodles_sam::record::data::{field::{Tag, Value}, Field};
-    /// let field = Field::new(Tag::AlignmentHitCount, Value::Int(1));
+    /// let field = Field::new(Tag::AlignmentHitCount, Value::Int32(1));
     /// assert_eq!(field.tag(), Tag::AlignmentHitCount);
     /// ```
     pub fn tag(&self) -> Tag {
@@ -48,8 +48,8 @@ impl Field {
     ///
     /// ```
     /// use noodles_sam::record::data::{field::{Tag, Value}, Field};
-    /// let field = Field::new(Tag::AlignmentHitCount, Value::Int(1));
-    /// assert!(matches!(field.value(), Value::Int(n) if *n == 1));
+    /// let field = Field::new(Tag::AlignmentHitCount, Value::Int32(1));
+    /// assert!(matches!(field.value(), Value::Int32(n) if *n == 1));
     /// ```
     pub fn value(&self) -> &Value {
         &self.value