+++ noodles-sam/src/record/data/field/value/ty.rs
@@ -12,7 +12,7 @@ pub enum Type {
     /// Character (`A`).
     Char,
     /// Integer (`i`).
-    Int,
+    Int32,
     /// Single-precision floating-point (`f`).
     Float,
     /// String (`Z`).
@@ -51,7 +51,7 @@ impl FromStr for Type {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
             "A" => Ok(Self::Char),
-            "i" => Ok(Self::Int),
+            "i" => Ok(Self::Int32),
             "f" => Ok(Self::Float),
             "Z" => Ok(Self::String),
             "H" => Ok(Self::Hex),
@@ -65,7 +65,7 @@ impl From<Type> for char {
     fn from(ty: Type) -> Self {
         match ty {
             Type::Char => 'A',
-            Type::Int => 'i',
+            Type::Int32 => 'i',
             Type::Float => 'f',
             Type::String => 'Z',
             Type::Hex => 'H',
@@ -81,7 +81,7 @@ mod tests {
     #[test]
     fn test_fmt() {
         assert_eq!(Type::Char.to_string(), "A");
-        assert_eq!(Type::Int.to_string(), "i");
+        assert_eq!(Type::Int32.to_string(), "i");
         assert_eq!(Type::Float.to_string(), "f");
         assert_eq!(Type::String.to_string(), "Z");
         assert_eq!(Type::Hex.to_string(), "H");
@@ -91,7 +91,7 @@ mod tests {
     #[test]
     fn test_from_str() {
         assert_eq!("A".parse(), Ok(Type::Char));
-        assert_eq!("i".parse(), Ok(Type::Int));
+        assert_eq!("i".parse(), Ok(Type::Int32));
         assert_eq!("f".parse(), Ok(Type::Float));
         assert_eq!("Z".parse(), Ok(Type::String));
         assert_eq!("H".parse(), Ok(Type::Hex));
@@ -108,7 +108,7 @@ mod tests {
     #[test]
     fn test_from_type_for_char() {
         assert_eq!(char::from(Type::Char), 'A');
-        assert_eq!(char::from(Type::Int), 'i');
+        assert_eq!(char::from(Type::Int32), 'i');
         assert_eq!(char::from(Type::Float), 'f');
         assert_eq!(char::from(Type::String), 'Z');
         assert_eq!(char::from(Type::Hex), 'H');